# contracts/openapi/wellbe-v2.openapi.json
{
  "components": {
    "schemas": {
...
      "ExportPacketResponse": {
        "properties": {
          "export_warning": {
            "default": "This is an exported copy. Once shared or saved by a recipient it cannot be recalled \u2014 revoking a link only stops future access.",
            "title": "Export Warning",
            "type": "string"
          },
          "packet": {
            "$ref": "#/components/schemas/VisitPacketV2"
          },
          "schema_version": {
            "const": "c13.visit_packet.export.v2",
            "default": "c13.visit_packet.export.v2",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "packet"
        ],
        "title": "ExportPacketResponse",
        "type": "object"
      },
      "GenerateVisitPacketRequest": {
        "properties": {
          "include_summary": {
            "default": true,
            "title": "Include Summary",
            "type": "boolean"
          },
          "prep": {
            "$ref": "#/components/schemas/PatientPrepInput"
          },
          "thread_ids": {
            "items": {
              "type": "string"
            },
            "title": "Thread Ids",
            "type": "array"
          },
          "time_window_end": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Window End"
          },
          "time_window_start": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Window Start"
          },
          "title": {
            "default": "Visit packet",
            "title": "Title",
            "type": "string"
          }
        },
        "title": "GenerateVisitPacketRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SharedPacketView": {
        "description": "Recipient-facing read of a shared packet (no owner-only internals).",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "review_note": {
            "default": "Patient-prepared and not clinician-reviewed. Every statement links to its source. This packet does not contain a diagnosis from WellBe.",
            "title": "Review Note",
            "type": "string"
          },
          "schema_version": {
            "const": "c13.visit_packet.shared_view.v2",
            "default": "c13.visit_packet.shared_view.v2",
            "title": "Schema Version",
            "type": "string"
          },
          "shared_by_label": {
            "default": "the patient",
            "title": "Shared By Label",
            "type": "string"
          },
          "statements": {
            "items": {
              "$ref": "#/components/schemas/VisitPacketStatementV2"
            },
            "title": "Statements",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "expires_at"
        ],
        "title": "SharedPacketView",
        "type": "object"
      },
...
      "UpdatePacketRequest": {
        "properties": {
          "inclusions": {
            "items": {
              "$ref": "#/components/schemas/UpdateStatementInclusion"
            },
            "title": "Inclusions",
            "type": "array"
          }
        },
        "title": "UpdatePacketRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Single external boundary. All surfaces call through here. See docs/architecture/component-map.md C13.",
    "title": "WellBe API (v2)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v2/share/{token}": {
      "get": {
        "description": "Public recipient read of a shared packet. Returns 404 for any unknown,\nrevoked, expired, or passcode-failing link (no information leak).",
        "operationId": "read_shared_packet_v2_share__token__get",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "title": "Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Share-Passcode",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Share-Passcode"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedPacketView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Shared Packet",
        "tags": [
          "v2-visit-packet"
        ]
      }
    },
...
    "/v2/visit-packets": {
      "post": {
        "operationId": "generate_visit_packet_v2_visit_packets_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-wellbe-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-patient-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Patient-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-type",
            "required": false,
            "schema": {
              "default": "controller",
              "title": "X-Wellbe-Actor-Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateVisitPacketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitPacketV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Visit Packet",
        "tags": [
          "v2-visit-packet"
        ]
      }
    },
    "/v2/visit-packets/{packet_id}": {
      "get": {
        "operationId": "get_visit_packet_v2_visit_packets__packet_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "packet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Packet Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-patient-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Patient-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-type",
            "required": false,
            "schema": {
              "default": "controller",
              "title": "X-Wellbe-Actor-Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitPacketV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Visit Packet",
        "tags": [
          "v2-visit-packet"
        ]
      },
      "patch": {
        "description": "Toggle statement inclusion. Deselected statements are kept and marked,\nnever silently dropped (decision: deselection visibility).",
        "operationId": "update_visit_packet_v2_visit_packets__packet_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "packet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Packet Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-patient-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Patient-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-type",
            "required": false,
            "schema": {
              "default": "controller",
              "title": "X-Wellbe-Actor-Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePacketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitPacketV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Visit Packet",
        "tags": [
          "v2-visit-packet"
        ]
      }
    },
    "/v2/visit-packets/{packet_id}/export": {
      "post": {
        "description": "Export a copy. This is a distinct, clearly-warned state from a controlled\nlink share: an exported copy cannot be recalled.",
        "operationId": "export_visit_packet_v2_visit_packets__packet_id__export_post",
        "parameters": [
          {
            "in": "path",
            "name": "packet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Packet Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-patient-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Patient-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-type",
            "required": false,
            "schema": {
              "default": "controller",
              "title": "X-Wellbe-Actor-Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportPacketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Visit Packet",
        "tags": [
          "v2-visit-packet"
        ]
      }
    },
    "/v2/visit-packets/{packet_id}/share": {
      "post": {
        "operationId": "share_visit_packet_v2_visit_packets__packet_id__share_post",
        "parameters": [
          {
            "in": "path",
            "name": "packet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Packet Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Actor-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-patient-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Wellbe-Patient-Id"
            }
          },
          {
            "in": "header",
            "name": "x-wellbe-actor-type",
            "required": false,
            "schema": {
              "default": "controller",
              "title": "X-Wellbe-Actor-Type",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-correlation-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          },
          {
            "in": "header",
            "name": "x-trace-id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SharePacketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharePacketResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Share Visit Packet",
        "tags": [
          "v2-visit-packet"
        ]
      }
    },
...
  }
}